+++ ofrak_patch_maker/CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 - `-fno-pic` flag added to the GNU_10_Toolchain to omit GOTs in patches (FEMs) against binaries that aren't dynamically linked. (see [#245](https://github.com/redballoonsecurity/ofrak/pull/245))
 
+### Added
+- Add methods to parse relocation symbols from object files.
+- Extend parsed symbol dictionary to include LinkableSymbolType.
+- Extend AssembledObject and BOM types to include relocation and unresolved symbols.
+
 ## [3.0.0](https://github.com/redballoonsecurity/ofrak/compare/ofrak-patch-maker-v.2.0.0...ofrak-patch-maker-v.3.0.0) - 2023-01-20
 ### Added
 - `-fno-optimize-sibling-calls` flag added to AVR toolchain.